% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{print.MCMC_DLM}
\alias{print.MCMC_DLM}
\title{Print Method for MCMC_DLM Object}
\usage{
\method{print}{MCMC_DLM}(x, n = 6, ...)
}
\arguments{
\item{x}{An MCMC_DLM object.}

\item{n}{Number of rows of the posterior to display.}

\item{...}{Further arguments; currently unused.}
}
\description{
Print output for MCMC_DLM objects. Displays basic information of regression model and the first few rows of the posterior.
}
\author{
Daniel Dempsey (\href{mailto:dempsed6@tcd.ie}{dempsed6@tcd.ie})
}
