% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{summary.MCMC_DLM}
\alias{summary.MCMC_DLM}
\title{Summary Method for MCMC_DLM Object}
\usage{
\method{summary}{MCMC_DLM}(object, digits = 2, ...)
}
\arguments{
\item{object}{An MCMC_DLM object.}

\item{digits}{Number of decimal places to round to for reported statistics.}

\item{...}{Further arguments; currently unused.}
}
\description{
Summary output for MCMC_DLM objects. Displays summary statistics of the posterior.
}
\author{
Daniel Dempsey (\href{mailto:dempsed6@tcd.ie}{dempsed6@tcd.ie})
}
