% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_posterior.R
\name{vis_output}
\alias{vis_output}
\title{Visualises Results from MCMC_DLM Fits}
\usage{
vis_output(x, type = "beta", include_intercept = FALSE, print_output = TRUE)
}
\arguments{
\item{x}{An MCMC_DLM object.}

\item{type}{One of "beta", "gamma" or "xi". Partial matching is supported. Each type represents a graph for a different parameter: beta gives a ridgeplot of the regression slopes, gamma plots the mean probability of inclusion, and xi (negative binomial regression only) displays a kernel density estimate of the negative binomial stopping parameter.}

\item{include_intercept}{Logical indicating if the intercept should be included in the graphs.}

\item{print_output}{Logical indicating whether or not the ggplot objects should be printed to the screen.}
}
\value{
A ggplot object. Will also display the ggplot visual if print_output = TRUE.
}
\description{
Provides some ggplot visualisations of the posterior from MCMC_DLM objects.
}
\author{
Daniel Dempsey (\href{mailto:dempsed6@tcd.ie}{dempsed6@tcd.ie})
}
