% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_DLM.R
\name{dataframe_DLM}
\alias{dataframe_DLM}
\title{Creating a Lagged Dataframe}
\usage{
dataframe_DLM(X, lag, dynamic_vars = NULL, arglag = list(fun = "bs"), ...)
}
\arguments{
\item{X}{A dataframe.}

\item{lag}{Lag length. See the help file for the "crossbasis" function from the dlnm package.}

\item{dynamic_vars}{The names of the columns that should be lagged. If not supplied, the dataset is not altered in any way (except possibly for scaling) and a warning is supplied.}

\item{arglag}{A list that is passed into onebasis for generating a basis matrix.}

\item{...}{Further arguments to be passed into the crossbasis function.}
}
\value{
A dataframe where the listed dynamic variables have been appropriately lagged. If no dynamic variables are given, the input dataframe is returned unaltered with a warning.
}
\description{
Transforms a dataframe by smoothing over lags of dynamic variables to prepare it for distributed lag modelling.
}
\examples{
X <- dplyr::select( dlnm::chicagoNMMAPS, c('cvd', 'dow', 'temp', 'dptp', 'o3') )
X <- na.omit( X )
arglag <- list( fun = 'bs', df = 4 )
DLM_dat <- dataframe_DLM( X, lag = 40, dynamic_vars =  c('temp', 'dptp', 'o3'), arglag = arglag )
}
\author{
Daniel Dempsey (\href{mailto:dempsed6@tcd.ie}{dempsed6@tcd.ie})
}
