% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{plot.MCMC_DLM}
\alias{plot.MCMC_DLM}
\title{Plotting Method for MCMC_DLM Object}
\usage{
\method{plot}{MCMC_DLM}(x, type = "beta", ...)
}
\arguments{
\item{x}{An MCMC_DLM object.}

\item{type}{One of "beta", "gamma" or "xi". Partial matching is supported. Each type represents a graph for a different parameter: beta gives a ridgeplot of the regression slopes, gamma plots the mean probability of inclusion, and xi (negative binomial regression only) displays a kernel density estimate of the negative binomial stopping parameter.}

\item{...}{Further arguments; currently unused.}
}
\description{
Provides some visualisations of the posterior.
}
\author{
Daniel Dempsey (\href{mailto:dempsed6@tcd.ie}{dempsed6@tcd.ie})
}
