% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QB_MCMC.R
\name{QB_MCMC}
\alias{QB_MCMC}
\title{Binary Quantile Regression via MCMC}
\usage{
QB_MCMC(
  formula,
  data = NULL,
  quantile = 0.5,
  nsamp = 5000,
  nburn = 5000,
  thin = 1,
  prior_beta_mu = 0,
  prior_beta_sigma = 100,
  prior_gamma_p = 0.5,
  init_beta = 0,
  init_gamma = FALSE
)
}
\arguments{
\item{formula}{Formula object to set the symbolic description of the model to be fitted.}

\item{data}{Optional dataframe, list or environment containing the model variables.}

\item{quantile}{The chosen quantile for regression.}

\item{nsamp}{The desired sample size from the posterior. Set to 5000 by default.}

\item{nburn}{The number of iterations of the MCMC to be discarded as burn-in. Set to 5000 by default.}

\item{thin}{Thinning factor of the MCMC chain after burn-in. Set to 1 by default (no values discarded after burn-in).}

\item{prior_beta_mu}{Mean of the Gaussian prior for the regression coefficients, beta. Either a vector of length equal to the number of predictors or a single numeric to represent a constant vector.}

\item{prior_beta_sigma}{Covariance matrix of the Gaussian prior for the regression coefficients, beta. Either a square matrix of dimension equal to the number of predictors or a single numeric to represent an isotropic covariance matrix.}

\item{prior_gamma_p}{Probability parameter of the Bernoulli prior for the predictor inclusion parameter, gamma. Either a vector of length equal to the number of predictors or a single numeric, to represent that all predictors have the same prior probability of inclusion. Note that the model must have one predictor included by default; because of this the first value must be equal to 1 if a vector is given.}

\item{init_beta}{Initial MCMC values for the beta parameters. Either a vector of length equal to the number of predictors or a single numeric representing the same starting value for each beta component.}

\item{init_gamma}{Initial MCMC values for the gamma parameters. Either a vector of length equal to the number of predictors or a single numeric representing the same starting value for each gamma component. Note that the model must have one predictor included by default; because of this the first value must be equal to 1 if a vector is given.}
}
\value{
A list containing the MCMC-derived posterior sample, as well as the data that were used.
}
\description{
Fits a quantile regression model to a binary response dataset using Gibbs sampling and performs Bayesian variable selection.
}
\details{
This function fits a quantile binary regression model via MCMC. Latent variable representation allows for Gibbs sampling of the parameters; see Benoit (2017). The algorithm also includes predictor inclusion uncertainty inference (inferred via a Metroplis step) adapted from Holmes and Held (2006). The parameters of interest for this model are the regression slopes (beta) and the binary predictor inclusion indicator (gamma).

As this is a Bayesian model, priors must be specified. Beta has a Gaussian prior and gamma has a Bernoulli prior.
}
\examples{
set.seed( 100 )
Pima_ex <- MASS::Pima.tr
Pima_ex$type <- ifelse( MASS::Pima.tr$type == 'Yes', 1, 0 )
pima_mcmc <- QB_MCMC( type ~ ., data = Pima_ex )
}
\references{
Dries F. Benoit and Dirk Van den Poel. "bayesQR: A Bayesian approach to quantile regression." Journal of Statistical Software 76 (2017): 1-32.

Chris C. Holmes and Leonhard Held. "Bayesian auxiliary variable models for binary and multinomial regression." (2006): 145-168.
}
\author{
Daniel Dempsey (\href{mailto:dempsed6@tcd.ie}{dempsed6@tcd.ie})
}
